Sanitation Update

+++ src/com/fratics/precis/sanitation/rules/SingleRuleResult.java
@@ -3,9 +3,9 @@ package com.fratics.precis.sanitation.rules;
 public class SingleRuleResult {
     public String ruleName;
     public boolean result;
-    public long fieldIndex;
+    public int fieldIndex;
 
-    public SingleRuleResult(String ruleName, boolean result, long fieldIndex) {
+    public SingleRuleResult(String ruleName, boolean result, int fieldIndex) {
 	this.ruleName = ruleName;
 	this.result = result;
 	this.fieldIndex = fieldIndex;
